% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{annotatedMatrixExampleFile}
\alias{annotatedMatrixExampleFile}
\title{Annotated Matrix Example File}
\usage{
annotatedMatrixExampleFile(file = "Symbol-To-Gene.mtx", list = FALSE,
  pkg = "AnnotatedMatrix")
}
\arguments{
\item{file}{Default \code{"Symbol-To-Gene.mtx"}. The file to
recover from inst/}

\item{list}{Default \code{FALSE}. If TRUE, then list all available
example files}

\item{pkg}{Default "AnnotatedMatrix", the R package to inspect
for the requested file.}
}
\value{
The path to the file requested, unless list is TRUE, which
    will return a listing of all files in \code{extdata/}
}
\description{
Path to a small example file used for demonstrations
}
\details{
This is just a wrapper around the internal .makeTempFile() function
that pulls up a gene symbol -to-> Entrez Gene ID mapping matrix
(subset of human genes). The file is factorized and has metadata on
both dimensions.
}
\examples{

ef1 <- annotatedMatrixExampleFile()
ef2 <- annotatedMatrixExampleFile("Machines.gmt")

}
