% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{annotatedMatrixParameter}
\alias{annotatedMatrixParameter}
\title{Annotated Matrix Parameter}
\usage{
annotatedMatrixParameter(key, value = NULL, default = NULL,
  as.list = NULL)
}
\arguments{
\item{key}{Required, a character vector of keys to recover. Keys
are case-insensitive, though the names on the returned values
will honor the capitalization of the passed `key` parameter.}

\item{value}{Optional new value to be assigned to the keys}

\item{default}{Default \code{NULL}. If the parameter is NULL, and
the corresponding R option is also NULL, this value will be
substitued instead}

\item{as.list}{Default \code{NULL}. If \code{TRUE}, the return
value will be a named list. \code{FALSE} will return the
unlist()ed values (which may be odd if the parameters have
mixed data types). The default of NULL will return a list if
two or more keys were requested, otherwise the single value}
}
\value{
If as.list=TRUE, a named list. If as.list=FALSE, a single
    value if a single key was requested, otherwise an unlist()ed
    vector.
}
\description{
Get/Set a parameter associated with your AnnotatedMatrix session
}
\details{
Manages values that you will likely want to be constant during an
analysis, but you may wish to customize between analyses, or
between users or organizations.

If a requested key is NULL, the function will also check if an R
option is set for that key. For example, if you request 'dir', the
option 'annotatedmatrixdir' will be checked if the parameter is
NULL.
}
\section{Recognized Keys}{


\itemize{

  \item \code{dir} - A default directory where matrix files are stored
  \item \code{ensemblhost} - Default Ensmebl host domain to use
}
}

