% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{appliedFilters}
\alias{appliedFilters}
\title{Applied Filters}
\arguments{
\item{new}{Default NULL, optional character vector of new filters
to apply. NOT YET IMPLEMENTED}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
The \link{setFilters} field, a character vector
    added to \link{setFilters}
}
\description{
AnnotatedMatrix object method to show/add applied filters
}
\details{
\preformatted{
## Method Usage:
myObject$appliedFilters( help=TRUE )

myObject$appliedFilters( new=NULL )
}

Each time a method is run that filters the matrix, the criteria
used are serialized as a text snippet, and the snippet is added to
the \link{setFilters} vector. These snippets can be used as a
record of filtering operations. In addition, some or all of them
can be passed to this function to re-run the filters

(LOADING = ASPIRATIONAL FEATURE, NOT YET AVAILABLE).
}
\examples{

s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
## The sample file has automated filters applied to it:
s2e$appliedFilters()
}
\seealso{
\link{setFilters}, \link{filterByScore},
    \link{filterByFactorLevel}, \link{filterByMetadata},
    \link{rNames}, \link{cNames}, \link{reset},
    \link{filterSummary},
}
