% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{as.gmt}
\alias{as.gmt}
\title{As GMT}
\arguments{
\item{obj}{Default NULL, the matrix object to serialize. If NULL,
will call \link{matObj} to recover the current filtered matrix.}

\item{transpose}{Default FALSE, which will treat rows as sets and
columns as set members. TRUE will instead treat the columns as
sets.}

\item{file}{Default NULL. If not NULL, will be treated as a file
path and passed to \link{cat}.}

\item{...}{Will be passed to \link{matObj}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
The GMT text, invisibly if file is not NULL
}
\description{
AnnotatedMatrix object method to represent matrix as GMT format
}
\details{
\preformatted{
## Method Usage:
myObject$as.gmt( help=TRUE )

myObject$as.gmt( obj=NULL, transpose=FALSE, file=NULL, ...)
}

Will represent the matrix as GMT format, a simple flat file format
that allows for basic metadata (names and descriptions).
}
\examples{

s2e  <- AnnotatedMatrix( annotatedMatrixExampleFile() )
## Generate GMT data for the symbols associated for each gene in
## the example:
s2e$as.gmt( transpose=TRUE )
}
\seealso{
\href{http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#GMT:_Gene_Matrix_Transposed_file_format_.28.2A.gmt.29}{GMT Format at Broad Institute}
}
