% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{as.ijx}
\alias{as.ijx}
\title{As IJX}
\arguments{
\item{file}{Default NULL. When NULL, a character matrix is
returned. If not NULL, the serialized data are written to the
specified file, as well as being invisibly returned.}

\item{sep}{Default \code{"\t"}, the field separator}

\item{obj}{Default \code{NULL}, will be passed to \code{melt}. If
\code{NULL}, then a character matrix will be returned,
otherwise a string representing the text is generated.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\description{
AnnotatedMatrix object method to convert matrix into three column format
}
\details{
This method is very similar to \code{\link{melt}}, which it calls
to structure the data. Unlike \code{melt}, it is primarily designed
to return a block of text (rather than a data.frame), or a
character matrix. It can also be used to write the text block to a
file.
}
\seealso{
\code{\link{melt}} to get a three column data.frame
}
