% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{as.mtx}
\alias{as.mtx}
\title{As MTX}
\arguments{
\item{obj}{Default NULL, the matrix object to serialize. If NULL,
will call \link{matObj} to recover the current filtered matrix.}

\item{file}{Default NULL. If not NULL, will be treated as a file
path and passed to \link{cat}.}

\item{\dots}{Will be passed to \link{matObj}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
The MTX text, invisibly if file is not NULL
}
\description{
AnnotatedMatrix object method to represent matrix as Matrix Market format
}
\details{
\preformatted{
## Method Usage:
myObject$as.mtx( help=TRUE )

myObject$as.mtx( obj=NULL, file=NULL, ...)
}

Will represent the matrix as Matrix Market format, a basic flat
file format that is well-suited for encoding sparse
matrices. Additionally, rich metadata will be added as comments
that can then be parsed by AnnotatedMatrix.

 #################################
 ### THIS FEATURE IS NOT YET CODED
 #################################
}
\examples{

s2e  <- AnnotatedMatrix( annotatedMatrixExampleFile() )
## Filter the matrix so that it is smaller:
s2e$rNames(c("HFH2","AIS1","p63"))
s2e$as.mtx()

}
\seealso{
\href{https://math.nist.gov/MatrixMarket/formats.html}{MTX Format at Broad Institute}
}
