% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{as.sidecar}
\alias{as.sidecar}
\title{As Metadata Sidecar}
\arguments{
\item{obj}{Default NULL, the matrix object to serialize. If NULL,
will call \link{matObj} to recover the current filtered matrix.}

\item{file}{Default NULL. If not NULL, will be treated as a file
path and passed to \link{cat}.}

\item{sep}{Default \code{"\t"}, the separator to use between
columns. The special value \code{NULL} will return a character
\code{matrix} object rather than a text representation.}

\item{\dots}{Will be passed to \link{matObj}. If you wish to
transpose the matrix, you can pass the argument here.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
If sep is \code{NULL}, a character matrix. Otherwise the
    text, invisibly if file is not NULL
}
\description{
AnnotatedMatrix object method to export metadata as a 'sidecar' file
}
\details{
\preformatted{
## Method Usage:
myObject$as.sidecar( help=TRUE )

myObject$as.sidecar( obj=NULL, file=NULL, sep="\t", ...)
}

 #################################
 ### THIS FEATURE IS NOT YET CODED
 #################################
}
