% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{biomartVersion}
\alias{biomartVersion}
\title{Current BioMart Version}
\usage{
biomartVersion(host = annotatedMatrixParameter("ensemblhost", default =
  "www.ensembl.org"), ...)
}
\arguments{
\item{host}{Will be passed to biomaRt, defaults to the
\link{annotatedMatrixParameter} "ensemblhost", falling back to
www.ensembl.org.}

\item{\dots}{Passed to \code{listMarts}. In particular, `host` may
be useful to access a preferred Ensembl server.}
}
\value{
\code{NA} if there is a problem, otherwise a single integer
    value representing the version, with a name representing the
    name of the Mart.
}
\description{
Report the current 'live' version of BioMart
}
\details{
Gets available Marts and looks for the one named
"Ensembl Genes ##", returning the actual value of '##' as the
version.
}
