% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{colDefs}
\alias{colDefs}
\title{Column Definitions}
\value{
A list, keyed to column names with descriptive text as
    values
}
\description{
Internal AnnotatedMatrix field holding metadata column descriptions
}
\details{
This list holds optional descriptive text for the metadata columns
associated with the matrix. It can be set 'automatically' by '%
ColumnDescription' comments in an MTX file, but in theory could be
'manually' set by code as well.

The descriptions will be shown as 'comments' when available
metadata columns are summarized, and will be included in the
"Columns" attribute of \link{map} output.
}
\examples{

s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
s2e$colDefs

# Definitions are also shown in summary text:
s2e

# They are also embedded in the attributes of $map() results:
myMap <- s2e$map('AIRE1')
attr(myMap, "Columns")[[ "Output Symbol" ]]
}
\seealso{
\link{matrixMD}, \link{map}
}
