% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{counts}
\alias{counts}
\alias{rCounts}
\alias{cCounts}
\title{Row and Column Counts}
\arguments{
\item{obj}{Default NULL, which will recover the matrix using
\link{matObj}. Optionally can be a user-provided sparse Matrix.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}

\item{...}{Passed to \link{matObj} (for example, \code{raw})}
}
\value{
A named vector of integers, which will be as long as the
    number of rows (for rCounts) or columns (for cCounts). Names
    are taken from the dimension names.
}
\description{
AnnotatedMatrix object methods to count non-zero entries for rows or columns
}
\details{
\preformatted{
## Method Usage:
myObject$rCounts( help=TRUE )
myObject$cCounts( help=TRUE )

myObject$rCounts( obj=NULL )
myObject$cCounts( obj=NULL )
}

For \code{$rCounts()}, return the counts of non-zero columns for
each row. Same for \code{$cCounts()}, which will return the counts
of non-zero rows for each column.

A value of zero indicates that the row (or column) does not have
any "connections" to the other dimension. A value of 1 indicates a
single unique connection.
}
\examples{

## Example matrix has gene symbols in rows, gene IDs in columns
s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
symCount <- s2e$rCounts()

# The matrix is loaded with an automatic filter. Which symbols no
# longer connect to any IDs?
names( symCount[ symCount == 0 ] )

# Which symbols are highly ambiguous (assigned to many genes)?
symCount[ symCount > 3 ]
# (Note that this matrix is case-sensitive, and includes "p40" and "P40")

idCount <- s2e$cCounts()
# Which genes have a lot of symbols assigned to them?
idCount[ idCount > 7 ]
# What are they?
s2e$map( names(idCount[ idCount > 7 ]) )
}
\seealso{
\link{populated}, \link{nnZero}
}
