% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{.filePattern}
\alias{.filePattern}
\title{File Pattern}
\usage{
.filePattern(qry, regexp = FALSE)
}
\arguments{
\item{qry}{Required, the query string}

\item{regexp}{Defaul \code{FALSE}, which will force a match on the
whole string. A value of TRUE will allow substring matches}
}
\description{
Build a query string for findMatrices that includes some wildcards
}
\keyword{internal}
