% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileLoaders.R
\name{.sidecarMatrixDimensions}
\alias{.sidecarMatrixDimensions}
\title{Parse sidecar dimensions}
\usage{
.sidecarMatrixDimensions(metadata, defRowDim = "RowItem",
  defColDim = "ColItem")
}
\description{
Parse row and column dimensions from sidecar information
}
\details{
When parsing metadata from sidecar files, parameters are smuggled
into an attribute of the 'main' metadata data.table. When
constructing an AnnotatedMatrix object, the 'RowDim' and 'ColDim'
parameters may be needed before the final processing in the object
method \code{$.readMatrix}. This function checks for those two
parameters and extracts them if present
}
\keyword{internal}
