% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{DOTaddAppliedFilter}
\alias{DOTaddAppliedFilter}
\alias{.addAppliedFilter}
\title{Add Applied Filter}
\arguments{
\item{key}{Required, a key indicating the filter type, eg 'SCORE'
or 'COUNT'}

\item{val}{Required, text defining the filter parameters}

\item{com}{Default NULL, optional comment (reason) for the filter}

\item{pre}{Default NULL, optional text (eg operator) to go before
the values}

\item{pro}{Default NULL, text (eg modifiers) to go after the values}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A single character string representing the fitler text
    added to \link{setFilters}
}
\description{
Internal AnnotatedMatrix object method to extend the list of applied filters
}
\details{
Internal method, should not be called directly

Helper method to add new filter entries to \link{setFilters}.
}
\seealso{
\link{filterLog}, \link{filterSummary},
    \link{appliedFilters}
}
