% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{DOTautoLevel}
\alias{DOTautoLevel}
\alias{.autoLevel}
\title{Auto Level}
\arguments{
\item{vals}{Required, vector of one or more factor levels}

\item{sep}{Default ',', the token used to separate multiple level
names}

\item{decreasing}{Default TRUE, which will sort factor values from
largest to smallest (in order to put the 'best'-scored factor
at the front of the list)}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
An integer representing the factor level
}
\description{
Internal AnnotatedMatrix object method to generate new hybrid factor levels
}
\details{
Internal method, should not be called directly

Takes a set of factor levels (character names), combines them via
paste, generates a new factor level if needed, and returns the
factor index. Will extend the \code{lvlVal} internal field as
needed.

This function is used by \link{map} to generate synthetic factor
levels when \code{collapse} causes multiple rows to be merged into
a single one.
}
\seealso{
\link{map}
}
