% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{DOTdetailZeroedRowCol}
\alias{DOTdetailZeroedRowCol}
\alias{.detailZeroedRowCol}
\title{Detail Zeroed Rows and Columns}
\description{
Internal AnnotatedMatrix object method to tally filtered rows and cols
}
\details{
Internal method, should not be called directly

Takes a SparseMatrix and a (previously calculated) logical vector
of 'failed' i+j triples, and determines which (if any) rows and
columns went from populated to unpopulated (all zeroes). Adds
entries to \link{filterLog} to record them
}
