% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{DOTfilterDetails}
\alias{DOTfilterDetails}
\alias{.filterDetails}
\title{Filter Details}
\arguments{
\item{id}{Required, a vector of one or more IDs (row or column
names) that have been filtered out.}

\item{type}{Optional type of filter, generally Row, Col or Cell}

\item{metric}{Optional human-readable criteria for the filter, such
as "Score > 3.14"}

\item{reason}{Optional human-readable rationale for why the filter
was applied, such as "Remove low-confidence assignments"}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
The \link{filterLog} filed (data.table) itself
}
\description{
Internal AnnotatedMatrix object method to add filtered names to the log
}
\details{
Internal method, should not be called directly

Helper method to add new filter entries to \link{filterLog}.
}
\seealso{
\link{filterLog}, \link{filterSummary},
    \link{appliedFilters}
}
