% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{DOTprocessMatrixData}
\alias{DOTprocessMatrixData}
\alias{.processMatrixData}
\title{Process Matrix Data}
\arguments{
\item{dat}{Required, a list of information related to the
matrix. The list can have the following components:
\itemize{
  \item matrix a sparse \code{Matrix} object
  \item colDefs Optional metadata column definitions
  \item metadata Optional data.table of metadata
  \item rowChanges Named vector of row names that had to be changed 
  \item colChanges Named vector of col names that had to be changed
  \item levels Optional level names for factorized matrices
  \item params Optional list of default parameters
}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
The list structure that aws provided by \code{dat}
}
\description{
Internal AnnotatedMatrix method to process structured matrix information
}
\details{
Internal method, should not be called directly. It is called by
both \link{.readMatrix} and \link{.readObject}. 

The method will determine the nature of the object:

At the moment, only \code{list} objects are recognized, which are
parsed with \link{matrixFromLists} and then processed with
\link{.processMatrixData}.
}
\seealso{
\link{.readMatrix}, \link{.readObject}
}
