% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{DOTreadMatrix}
\alias{DOTreadMatrix}
\alias{.readMatrix}
\title{Read Matrix Data}
\arguments{
\item{format}{Default "" (empty string). Optional format name, if
empty the format will be deduced by the file suffix.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A list containing the \link{Matrix} object, metadata as a
    \link{data.table}, \link{rowChanges} and \link{colChanges},
    levels as \link{lvlVal}, and any default parameters parsed from
    the file.
}
\description{
Internal AnnotatedMatrix object method to parse file contents
}
\details{
Internal method, should not be called directly

This method will take the \link{file} field as a file path and
attempt to parse it as a matrix. Once parsed, the method will
create a serailized \link[=readRDS]{RDS} object. Future attempts to
parse the matrix file will recognize the RDS object and quickly
parse it instead.
}
\seealso{
The individual format parsing functions:
    \link{parse_MatrixMarket_file}, link{filesToMTX},
    link{parse_ListOfLists_file}, link{parse_Text_file},
    link{parse_GMT_file}
}
