% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{DOTreadObject}
\alias{DOTreadObject}
\alias{.readObject}
\title{Read R Object}
\arguments{
\item{obj}{Required, the R object holding the matrix information.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A list structure with various components needed by the object.
}
\description{
Internal AnnotatedMatrix method to parse an R object as a matrix
}
\details{
Internal method, should not be called directly. It is called
automatically when a new object is created and an R object is
provided instead of a file.

The method will determine the nature of the object:

At the moment, only \code{list} objects are recognized, which are
parsed with \link{matrixFromLists} and then processed with
\link{.processMatrixData}.
}
\seealso{
\link{matrixFromLists}, \link{.processMatrixData}
}
