% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{file}
\alias{file}
\title{Matrix File Path}
\value{
A character vector with a single value
}
\description{
Internal AnnotatedMatrix field holding the path to the matrix file
}
\details{
This is simply the path to the raw file (or directory of files)
that define the matrix.
}
\examples{

s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
s2e$file

}
\seealso{
\link{fromRDS}, \link{.readMatrix}
}
