% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{filterByFactorLevel}
\alias{filterByFactorLevel}
\title{Filter by Factor Level}
\arguments{
\item{x}{Required, a list of factor levels. Can be either the
factor level names (character) or integer level assignments.}

\item{keep}{Default TRUE, which will presume that \code{x}
represents levels that should be kept. If FALSE, then levels in
\code{x} will be removed.}

\item{ignore.case}{Default TRUE, which will match factor levels
regardless of case.}

\item{filterEmpty}{Default FALSE. If true, then rows or columns
that are empty after filtering will be removed from the matrix
using \link{removeEmpty}. If no cells were affected by the
\code{filterByFactorLevel} call, then \code{removeEmpty} will not be
called. If you wish to assure that empty rows and columns are
removed after this filter, you should call \code{removeEmpty}
explicitly.}

\item{reason}{Default NA. Optional human-readable reason for why
the alteration was made, will be recorded in \link{filterLog}
and used to structure \link{filterSummary}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
Invisibly, an integer vector of newly zeroed counts for
    (cells, rows, cols).
}
\description{
AnnotatedMatrix object method to filter cells by factor level assignment
}
\details{
\preformatted{
## Method Usage:
myObject$filterByFactorLevel( help=TRUE )

myObject$filterByFactorLevel(x, keep=TRUE, ignore.case=TRUE,
                             filterEmpty=FALSE)
}

Updates the \link[=matrixUse]{used matrix} to remove (zero-out)
cells that do not have the requested factor levels assinged. Like
all filters, will not alter \link{matrixRaw}.

If the matrix is not factorized, an error will be reported, no
action will be taken, and NA will be invisibly returned.
}
\examples{

# The example matrix has some highly atypical symbols. To look at
# them, we will need to revert filters, since the file has
# autofilters that exclude them from the start:
s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
s2e$reset()

# Check the levels to make sure we recall what we want:
s2e$levels()

s2e$filterByFactorLevel(c("Unknown"), keep=TRUE, filterEmpty=TRUE,
                        reason="Inspecting odd symbols")
s2e$filterSummary()

# We set filterEmpty to TRUE in order to more easily inspect the
# matrix iteslf:
s2e$matrixUse
}
\seealso{
\link{filterLog}, \link{filterSummary},
    \link{appliedFilters}
}
