% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{filterByScore}
\alias{filterByScore}
\title{Filter by Score}
\arguments{
\item{min}{Optional minimum score. Cells that are below this value
will be set to zero}

\item{max}{Optional maximum score. Cells that are above this value
will be set to zero}

\item{filterEmpty}{Default FALSE. If true, then rows or columns
that are empty after filtering will be removed from the matrix
using \link{removeEmpty}. If no cells were affected by the
\code{filterByScore} call, then \code{removeEmpty} will not be
called. If you wish to assure that empty rows and columns are
removed after this filter, you should call \code{removeEmpty}
explicitly.}

\item{reason}{Default NA. Optional human-readable reason for why
the alteration was made, will be recorded in \link{filterLog}
and used to structure \link{filterSummary}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
Invisibly, an integer vector of newly zeroed counts for
    (cells, rows, cols).
}
\description{
AnnotatedMatrix object method to filter cells by their numeric value
}
\details{
\preformatted{
## Method Usage:
myObject$filterByScore( help=TRUE )

myObject$filterByScore( min=NA, max=NA, filterEmpty=FALSE, reason=NA )
}

Updates the \link[=matrixUse]{used matrix} to remove (zero-out)
cells that fail a numeric test. Like all filters, will not alter
\link{matrixRaw}.
}
\examples{

# The example matrix is factorized, but we can still fitler it by
# the numeric factor levels
s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )

# Check the levels:
s2e$levels()

# How many populated rows currently?
sum( s2e$populatedRows() )  # 151

# Let's only keep official symbols
s2e$filterByScore( min=4, reason="Focus on official symbols only" )
# We now have a more constrained matrix:
sum( s2e$populatedRows() )  # 28
s2e$filterSummary()

}
\seealso{
\link{filterLog}, \link{filterSummary},
    \link{appliedFilters}
}
