% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{filterLog}
\alias{filterLog}
\title{Filter Log}
\value{
A data.table
}
\description{
Internal AnnotatedMatrix field holding a data.table of rows or
columns removed by filters
}
\details{
This AnnotatedMatrix object field is \code{NULL} if no filters have
been applied. Otherwise it is a filtered form of
\link{matrixRaw}. You will generally not want to access this field
directly; Instead, use \link{matObj}, as it will by default provide
matrixUse if filters are applied, or will fallback to
matrixRaw. Using \link{reset} will reset all filters by setting
matrixUse to \code{NULL}.
}
\examples{

# The example matrix includes some automatic filters:
s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
s2e$filterSummary()  # Overview of filter effects
# What columns have been eliminated?
s2e$filterLog[ s2e$filterLog$type == 'Col' , c("id", "metric") ]

# Frank says he can't find data for p51 anymore. Was it filtered?
s2e$filterLog[ s2e$filterLog$id == "p51", ]
# Yes; Provide Frank with guidance on avoiding sketchy gene symbols

}
\seealso{
\link{reset}, \link{filterSummary}
}
