% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{filterSummary}
\alias{filterSummary}
\title{Filter Summary}
\arguments{
\item{byreason}{Default TRUE, which will cluster the filter summary
by reason and type. If FALSE, then only type will be used to
cluster the summary.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A data.frame classed as 'mapFilter'. Will be pretty-printed
    with \link{print.mapFilter}.
}
\description{
AnnotatedMatrix object method to summarize filters applied to the matrix
}
\details{
\preformatted{
## Method Usage:
myObject$filterSummary( help=TRUE )

myObject$filterSummary( reason=TRUE )
}

Returns a data.frame that has been classed to allow a pretty-print
summary of the filters applied and the number of rows or columns
that have been 'removed' by each filter.
}
\examples{

s2e  <- AnnotatedMatrix( annotatedMatrixExampleFile() )
fs <- s2e$filterSummary( )
fs
str(fs)
}
\seealso{
\link{setFilters}, \link{filterByScore},
    \link{filterByFactorLevel}, \link{filterByMetadata},
    \link{rNames}, \link{cNames}, \link{reset},
    \link{appliedFilters},
}
