% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{findMatrices}
\alias{findMatrices}
\title{Find Matrices}
\usage{
findMatrices(ns1 = NULL, ns2 = NULL, mod = NULL, type = NULL,
  auth = NULL, vers = NULL, dir = getOption("annotatedmatrixdir"),
  recursive = TRUE, ignore.case = TRUE, most.recent = TRUE,
  regexp = FALSE, clear.cache = FALSE)
}
\arguments{
\item{ns1}{Default \code{NULL}. Optional, the "first" namespace
(but will be matched to either the first or second namespace
field)}

\item{ns2}{Default \code{NULL}. Options, the "second" namespace,
will be matched to the other field relative to \code{ns1}}

\item{mod}{Default \code{NULL}. A modifier assigned to some
matrices, often a species name like "Human" or "Mouse".}

\item{type}{Default \code{NULL}. The type of the matrix, generally
"Map" or "Ontology"}

\item{auth}{Default \code{NULL}. The authority that provided the
data used to make the matrix, for example "Entrez",
"GeneOntology", "PubMed" etc}

\item{vers}{Default \code{NULL}. The version of the matrix, eg
"2017-11-05", "GRCh37", etc}

\item{dir}{Default
\code{getOption("annotatedmatrixdir")}. Required, the folder to
search in on your file system.}

\item{recursive}{Default \code{TRUE}, should subdirectories be
searched as well.}

\item{ignore.case}{Default \code{TRUE}. Should matches ignore case?}

\item{most.recent}{Default \code{NULL}. If NULL, keep all
results. Otherwise, cluster the matrices by all fields except
version, and keep only the "most recent" version of each
cluster. If \code{most.recent} contains the substring "vers",
this will be done by sorting the Version field of the
data.frame. Otherwise, the Modified field (file system
modification time) will be used.}

\item{regexp}{Default \code{FALSE}. Should matches be performed by
regular expression?}

\item{clear.cache}{Default \code{FALSE}. File lists are cached for
each \code{dir} requested (as long as \code{recursive} is
TRUE). This allows faster recovery of subsequent searches, but
will prevent discovery of newly-created matrices. Pass a value
of TRUE to assure an explicit search of the directory.}
}
\value{
A data.frame with the following fields:

\itemize{
  \item SubDir - subdirectory relative to \code{dir}
  \item Type - The type of conversion matrix
  \item Modifier - A modifier, generally a species
  \item NS1 - The row namespace / database
  \item NS2 - The column namespace / database
  \item Authority - The authority for the underlying data
  \item Version - The version of the matrix
  \item Path - The path to the matrix file, relative to \code{dir}
  \item Modified - The file modification time
}
}
\description{
Searches a directory for matrices matching your criteria
}
\details{
The expected file format will be one of:

\code{<TYPE>@<MOD>-<NS>_to_<NS>@<AUTH>@<VERS>.mtx}
\code{<TYPE>@<NS>_to_<NS>@<AUTH>@<VERS>.mtx}

... keeping in mind that the relative order of NS1 and NS2 is
irrelevant, and noting that some matricies will not have a modifier
(MOD). It will be expected that the suffix be lower case.
}
