% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{fromRDS}
\alias{fromRDS}
\title{Matrix Source RDS Flag}
\value{
A character vector with a single value
}
\description{
Internal AnnotatedMatrix field defining if loading was from a cached RDS
}
\details{
A logical flag, if TRUE it indicates that the matrix was loaded
from a cached RDS object. The RDS file will always be the same as
\link{file}, but with an additional '.rds' suffix appended.

If the RDS file does not already exist, \link{.readMatrix} will
parse the 'original' files and automatically create a new RDS
file. Similarly, if the RDS file is older than the original
sources, it will be regenerated from source.

\preformatted{
## NORMALLY YOU WILL NOT WANT TO ACCESS THIS FIELD DIRECTLY
}
}
\examples{

s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
s2e$fromRDS

}
\seealso{
\link{file}, \link{.readMatrix}
}
