% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{generateMtxRds}
\alias{generateMtxRds}
\title{Generate MTX RDS}
\usage{
generateMtxRds(path, recursive = TRUE)
}
\arguments{
\item{path}{Required, the path to the folder to inspect}

\item{recursive}{Default \code{TRUE}, which will cause subfolders
to also be processed}
}
\value{
A character vector of file paths that were processed
}
\description{
Little tool function to pre-calculate RDS files from MTX
}
\details{
Matrix generation scripts will make a collection of 'raw' (text)
MTX files. When these files are passed to AnnotatedMatrix, they
will be parsed to extract the sparse matrix, and to process the
various annotations. Since this takes time, the end result will be
cached as a serialized RDS object with the same file path plus
'.rds'.

This function merely takes a folder path as input, finds all .mtx
files under it, and loads them in AnnotatedMatrix to pre-generate
the RDS files. In some circumstances this may increase the
satisfaction of particularly impatient colleagues.
}
