% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{is.factor}
\alias{is.factor}
\title{Matrix is Factor}
\arguments{
\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A single logical value, TRUE if \link{lvlVal} is defined
    with one or more levels, otherwise FALSE.
}
\description{
AnnotatedMatrix object method indicating if the matrix is a factor or not
}
\details{
\preformatted{
## Method Usage:
myObject$is.factor( help=TRUE )

myObject$is.factor( )
}

\link{Matrix} matrices do not support factorized content, at least
at the time this package was developed. AnnotatedMatrix maintains
some metadata to allow integer matrices to be treated as factors
with named levels. This method will indicate if the \link{lvlVal}
field has ben set.
}
\examples{

s2e  <- AnnotatedMatrix( annotatedMatrixExampleFile() )
s2e$is.factor( )
s2e$levels()
}
\seealso{
\link{lvlVal}, \link{levels}
}
