% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{levels}
\alias{levels}
\title{Matrix Levels}
\arguments{
\item{asFactor}{Default FALSE, in which case factor level names
only will be returned. If TRUE, then the returned value will be
a formal factor object.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
NULL if the matrix is not a factor, or a character vector
    of level names, or a factor if \code{asFactor} is TRUE.
}
\description{
AnnotatedMatrix object method returning the factor levels (names)
}
\details{
\preformatted{
## Method Usage:
myObject$levels( help=TRUE )

myObject$levels( )
}

If the matrix is flagged as being a factor (by having \link{lvlVal}
set), this function will return the factor levels (names).
}
\examples{

s2e  <- AnnotatedMatrix( annotatedMatrixExampleFile() )
s2e$levels()
str( s2e$levels( asFactor=TRUE ) )
}
\seealso{
\link{lvlVal}, \link{is.factor}
}
