% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{lvlVal}
\alias{lvlVal}
\title{Matrix Level Values}
\value{
A character vector of level values
}
\description{
Internal AnnotatedMatrix field holding factor levels (names)
}
\details{
AnnotatedMatrix objects can be defined as a "pseudo-factor" by
setting factor level names that will be associated with numeric
values of the matrix. These level names are held in the
\code{lvlVal} field. If the field is populated the matrix will be
presumed to be a factor.

\preformatted{
## THIS FIELD SHOULD NOT BE MODIFIED BY THE USER OR OTHER CODE
## DOING SO COULD LEAD TO CONFUSION / CRASHES
}
}
\examples{

s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
s2e$lvlVal
}
\seealso{
\link{is.factor}, \link{levels}
}
