% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEnsemblMatrix.R
\name{makeEnsemblMatrix}
\alias{makeEnsemblMatrix}
\title{Make Ensembl Matrix}
\usage{
makeEnsemblMatrix(dataset = NULL, ...)
}
\arguments{
\item{dataset}{Default \code{NULL}, which will recover all
available datasets from BioMart. Can alternatively be a vector
of one or more datasets.}

\item{\dots}{Passed on to BioMart functions}
}
\description{
Generates an Ensembl AnnotatedMatrix file
}
\details{
WORK IN PROGRESS
}
