% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{matObj}
\alias{matObj}
\title{Get Matrix Object}
\arguments{
\item{raw}{Default \code{FALSE}, which will return \link{matrixUse}
if filters have been applied. If no filters are applied, or
raw=TRUE, then \link{matrixRaw} will be returned}

\item{transpose}{Default \code{FALSE}. if TRUE, then transpose the
matrix before returning it.}
}
\value{
a Matrix:: 'dgTMatrix' sparse matrix
}
\description{
Recover the 'current' matrix data
}
\examples{

s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
## Automatically applied filters result in fewer non-zero elements
## of the filtered matrix than the raw one:
Matrix::nnzero(s2e$matObj())         # 180
Matrix::nnzero(s2e$matObj(raw=TRUE)) # 208

}
\seealso{
\link{reset}, \link{matrixRaw}, \link{matrixUse}
}
