% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileLoaders.R
\name{matrixFromLists}
\alias{matrixFromLists}
\title{Matrix from Lists}
\usage{
matrixFromLists(data, meta = NULL, listNames = NULL, file = NULL,
  val = 1)
}
\arguments{
\item{data}{Required, a list of vectors. Each list element is
considered to be a row, and the members will become (in
aggreagate) the columns}

\item{meta}{Default NULL. Optional list of named character vectors
representing metadata assignements. The list names represent
the metadata names (eg "Description", "Notes", whatever). The
vector names represent the thing being annotated (a row or
column name), and the values hold the actual
assignment. Metadata is "mixed" in that there are not separate
tables for columns and rows.}

\item{listNames}{Default NULL, in which case the listNames
(rNames) will be taken as names(data). The option to provide
the names independently is included in case some names are not
unique. They will still be made so in the final matrix}

\item{file}{Default NULL. If non-null, will be used to look for
sidecar metadata files.}

\item{val}{Default 1. The value to be assigned to non-zero cells in
the matrix.}
}
\value{
A list with the following members:

\itemize{

  \item mat - The sparse matrix

  \item metadata - data.table of any metadata assignments

  \item colChanges - a named character vector. The names are the
  column names as used in the matrix, the values are the names as
  originally provided. Will only contain names that had to be
  altered to avoid uniqueness conflicts.

  \item rowChanges - as above, but for rows

}

This structure can be read by internal methods used by
AnnotatedMatrix to parse flat files
}
\description{
Convert a list of vectors into a sparse matrix consumable by AnnotatedMatrix
}
\details{
Sparse matrices (eg dgTMatrix) are constructed from lists of
triples; row number (i), column number (j) and value of that cell
(x). This function generates such triples by providing one or more
lists of "members" (strings)
}
\examples{

ingredients <- list(Potato=c("Mashed","Fried","Baked"),
                    Onion=c("Fried"),
                    Apple=c("Raw","Baked"))

meta <- list(Type=setNames(c("Vegetable","Fruit","Vegetable"),
                           c("Potato","Apple","Onion")),
             Tool=setNames(c("Tray","Pan","Pot",NA),
                           c("Baked","Fried","Mashed","Raw")),
             Appliance=setNames(c("Oven","Stove","Stove",NA),
                                c("Baked","Fried","Mashed","Raw")))
                             
matrixFromLists(ingredients, meta)

}
