% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{matrixMD}
\alias{matrixMD}
\title{Matrix Metadata Object}
\value{
a data.table object
}
\description{
Internal AnnotatedMatrix field holding metadata about row and column entries
}
\details{
This field stores metadata information for the AnnotatedMatrix
object. Metadata for rows and columns are both stored in a single
\code{data.table} (this may be a bad idea - I may split the field
later into matrixMdRow and matrixMdCol)

\preformatted{
## NORMALLY YOU WILL NOT WANT TO ACCESS THIS FIELD DIRECTLY
## Use methods instead, in this case $metadata()
}
}
\examples{

s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
str(s2e$matrixMD)

}
\seealso{
\link{metadata}
}
