% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{matrixText}
\alias{matrixText}
\title{Matrix Text}
\arguments{
\item{pad}{Default \code{""}, text to prefix the object. Intended
to allow nested construction of the summary, such that multiple
AnnotatedMatrix objects embedded in a more complex object (eg
SetFisher) can be visually combined.}

\item{useObj}{Default \code{NULL}, the Matrix object to
summarize. When NULL \code{$matObj()} will be used.}

\item{fallbackVar}{Default \code{NULL}, a variable name to use when
\code{.selfVarName()} fails to extract the actual variable
name.}

\item{compact}{Default \code{FALSE}. If TRUE, then a more compact
report (fewer details) is produced.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A single string holding the text to be displayed
}
\description{
Generate text summary of the AnnotatedMatrix object
}
\details{
\preformatted{
## Method Usage:
myObject$matrixText( help=TRUE )

myObject$matrixText(pad = "", useObj=NULL, fallbackVar=NULL,
                    compact=FALSE, color=NULL)
}

All RefClass objects include a \code{show()} method, which is
similar to \code{print} in other classes - it will generate output
when the object is evaluated "as-is". AnnotatedMatrix uses
\code{matrixText} to assemble the text shown by show().
}
\examples{

s2e  <- AnnotatedMatrix( annotatedMatrixExampleFile() )
s2e$matrixText()         # just the raw text, with ANSI codes
cat( s2e$matrixText() )  # ... print it pretty, which is the same as:
s2e$show()               # ... which is the same as:
s2e                      # ... how $show() is "meant to be used"
}
