% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{matrixUse}
\alias{matrixUse}
\title{Used (Filtered) Matrix}
\value{
a Matrix:: 'dgTMatrix' sparse matrix, or NULL if no filters
    have been applied
}
\description{
Internal AnnotatedMatrix field holding the filtered Matrix object
}
\details{
This AnnotatedMatrix object field is \code{NULL} if no filters have
been applied. Otherwise it is a filtered form of
\link{matrixRaw}. You will generally not want to access this field
directly; Instead, use \link{matObj}, as it will by default provide
matrixUse if filters are applied, or will fallback to
matrixRaw. Using \link{reset} will reset all filters by setting
matrixUse to \code{NULL}.

\preformatted{
## NORMALLY YOU WILL NOT WANT TO ACCESS THIS FIELD DIRECTLY
## Use methods instead, in this case $matObj()
## ALTERING THIS FIELD WILL LEAD TO CHAOS. PLEASE DON'T.
}
}
\examples{

s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
# Example matrix has an autofilter, so matrixUse is populated:
str(s2e$matrixUse)
# Resetting filters will also clear $matrixUse, until new filters
# are applied.
s2e$reset()
str(s2e$matrixUse)

}
\seealso{
\link{matObj}, \link{matrixRaw}, \link{reset}
}
