% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{melt}
\alias{melt}
\title{Melt}
\arguments{
\item{obj}{Default NULL, the matrix object to serialize. If NULL,
will call \link{matObj} to recover the current filtered matrix.}

\item{file}{Default NULL. When NULL, the data.frame is simply
returned. If not NULL, the table is written to the specified
file, and the data.frame is invisibly returned.}

\item{named.dims}{Default \code{TRUE}. If FALSE, then the table
will have a column header of "Row","Col","Value". When TRUE,
Row and Col will be replaced by the relevant dimnames for the
matrix, if available. If a 'CellDim' parameter has been set, it
will replace Value.}

\item{\dots}{Will be passed to \link{matObj}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A data.frame, invisibly if \code{file} is specified
}
\description{
AnnotatedMatrix object method to convert matrix into three column table
}
\details{
\preformatted{
## Method Usage:
myObject$melt( help=TRUE )

myObject$melt( obj=NULL, file=NULL, named.dims=TRUE, ...)
}

Will represent the matrix as a "melted" table with three columns;
row name, column name, value. Every non-zero cell in the matrix
will be represented as a row. Zero-values are excluded.

The return value is a data.frame, which can be optionally written
to a file.
}
\examples{

s2e  <- AnnotatedMatrix( annotatedMatrixExampleFile() )
## Strip down to a smaller size for easier illustration
s2e$rNames(c("HFH2","AIS1","p63"))
s2e$melt( )

}
\seealso{
\code{\link{as.ijx}} to generate a matrix or a block of text.
}
