% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{metadata}
\alias{metadata}
\title{Metadata Keys}
\arguments{
\item{id}{Default NULL. Optional vector of specific IDs to query}

\item{key}{Default NULL. Optional vector of keys / tags to query}

\item{na.rm}{Default TRUE, which will cause rows or columns that
are entirely NA to be removed.}

\item{drop}{Default TRUE, which will cause a named vector to be
returned if only one column of metadata is present. Otherwise a
data.table will be returned.}

\item{verbose}{Default TRUE, which will warn about certain issues}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
If drop=TRUE and zero or one columns of metadata a present,
    a named character vector. Otherwise a data.table
}
\description{
AnnotatedMatrix object method to recover metadata assigned to rows or columns
}
\details{
\preformatted{
## Method Usage:
myObject$metadata( help=TRUE )

myObject$metadata(id=NULL, key=NULL, na.rm=TRUE, drop=TRUE, verbose=TRUE )
}

Query metadata associated with rows and/or columns, based on IDs
and/or keys
}
\examples{

s2e  <- AnnotatedMatrix( annotatedMatrixExampleFile() )
s2e$metadata(c("LOC9588","LOC7038","LOC3921"))
s2e$metadata(key="Symbol")
}
\seealso{
\link{metadata_keys}, \link{matrixMD}
}
