% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{metadata_keys}
\alias{metadata_keys}
\title{Metadata Keys}
\arguments{
\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A character vector of metadata column names
}
\description{
AnnotatedMatrix object method to return all metadata keys
}
\details{
\preformatted{
## Method Usage:
myObject$metadata_keys( help=TRUE )

myObject$metadata_keys( )
}

Return all keys (tag names) from the internal metadata table. The
\code{id} column will be excluded from the result.
}
\examples{

s2e  <- AnnotatedMatrix( annotatedMatrixExampleFile() )
## Two keys assigned to the example matrix:
s2e$metadata_keys()
}
\seealso{
\link{metadata}, \link{matrixMD}
}
