% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{nnZero}
\alias{nnZero}
\title{Number of Non-zero Cells}
\arguments{
\item{obj}{Default NULL, which will recover the matrix using
\link{matObj}. Optionally can be a user-provided sparse Matrix.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}

\item{...}{Passed to \link{matObj} (for example, \code{raw})}
}
\value{
An integer
}
\description{
AnnotatedMatrix object method to return a count of non-zero cells
in the used matrix
}
\details{
\preformatted{
## Method Usage:
myObject$nnZero( help=TRUE )

myObject$nnZero( obj=NULL )
}

For sparse matrices, the non-zero values represent the data, which
in AnnotatedMatrix are typically used to indicate connections of
some sort between the two dimensions of the matrix.
}
\examples{

s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
s2e
# The example matrix begins filtered - inspecting the object should show:

# 180 non-zero cells (1.92\%)
# [Raw Matrix] : 167 x 67 (90.4\%x92.5\%), 208 non-zero (1.86\%; 86.54\% survive filters)

# That is, in the filtered state 180 cells are non-zero. Confirm:
s2e$nnZero()           # 180
# We can see the total 'raw' connections by passing the raw argument:
s2e$nnZero( raw=TRUE ) # 208
# Or we can reset the filters and check again:
s2e$reset()
s2e$nnZero()           # 208
}
\seealso{
\link{counts}, \link{populated}
}
