% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{normalizePercent}
\alias{normalizePercent}
\title{Normalize percent}
\usage{
normalizePercent(x, denom = as.numeric(NA))
}
\arguments{
\item{x}{Required, a character or numeric vector containing the
values to be evaluated}

\item{denom}{Default NA. The "denominator" for the calculations,
which will be used when a percentage is provided in x.}
}
\description{
Convert text percentages (eg "4%") to numbers based on a denominator
}
\details{
Used to parse parameters that can be absolutely defined (eg 53) or
relatively defined as a percentage (eg "2.5%"). For the latter will
require a denominator to be provided as well
}
\examples{

AnnotatedMatrix:::normalizePercent( c("5", "0.02", "5\%", "0.02\%"), 1000)
AnnotatedMatrix:::normalizePercent( 1:10 ) # Numbers are 'left alone'

}
