% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileLoaders.R
\name{parse_GMT_file}
\alias{parse_GMT_file}
\title{Parse GMT File}
\usage{
parse_GMT_file(file)
}
\arguments{
\item{file}{Required, the path to the GMT file}
}
\value{
A list structure usable by AnnotatedMatrix
}
\description{
Parses one or more lists from a Gene Matrix Transposed file
}
\details{
GMT files allow an arbitrary number of lists to be stored, each
with a name, description, and one or more members

Generally you would not call this function directly.
}
