% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileLoaders.R
\name{parse_IJX_file}
\alias{parse_IJX_file}
\title{Parse IJX File}
\usage{
parse_IJX_file(file, sep = "\\t")
}
\arguments{
\item{file}{Required, the path to the IJX file}

\item{sep}{Default \code{"\t"}, the field separtor in the input
file.}
}
\value{
A list structure usable by AnnotatedMatrix
}
\description{
Parses one or more lists from a three column "ijx" (row col val) file
}
\details{
This format can be described as one row per cell. Each row
represents the row identifier (first column, i) the column
identifier (second column, j) and the value assigned to that
row/col cell (third column, x).

Generally you would not call this function directly.
}
