% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileLoaders.R
\name{parse_MatrixMarket_file}
\alias{parse_MatrixMarket_file}
\title{Parse MatrixMarket File}
\usage{
parse_MatrixMarket_file(file)
}
\arguments{
\item{file}{Required - the path to the .mtx file. Files can be
gzipped, provided they have a .gz suffix.}
}
\description{
Parses a MatrixMarket file so it can be read by AnnotatedMatrix
}
\details{
Reads the "normal" matrix data from a MatrixMarket file (the
[i,j,x] triples), as well as parses '%' comments to extract
metadata usable by AnnotatedMatrix.
}
