% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileLoaders.R
\name{parse_Text_file}
\alias{parse_Text_file}
\title{Parse Text File}
\usage{
parse_Text_file(files, header = FALSE, rm.suffix = TRUE, pattern = NULL)
}
\arguments{
\item{files}{Required. One or more file/directory paths (can be a
mixture of directories and files). Files can be gzipped,
provided they have a .gz suffix.}

\item{header}{Default FALSE. If true, the first row of each file
will be taken as the list name for that list}

\item{rm.suffix}{Default TRUE, which will remove the file suffix
when the file name is used as a list name}

\item{pattern}{Default NULL. Optional pattern to use when
extracting files from requested directories}
}
\value{
A list structure usable by AnnotatedMatrix
}
\description{
Parses one or more simple text files into data consumable by AnnotatedMatrix
}
\details{
Provided one or more files and/or directories, take all files and
treat each as a single list

Generally you would not call this function directly.
}
