% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileLoaders.R
\name{parseMetadataSidecar}
\alias{parseMetadataSidecar}
\title{Parse Metadata Sidecar}
\usage{
parseMetadataSidecar(file, metadata = NULL, na.strings = c("NA", "-"),
  verbose = TRUE)
}
\arguments{
\item{file}{Required, the path to the primary data file (NOT to the
metadata file)}

\item{metadata}{Default NULL. Optional data.table that already
contains metadata, presumably from another source. If not
provided, an empty data.table will be made.}

\item{na.strings}{Default \code{c("NA",'-')}, strings that will be
parsed as NA. A dash is included as it is a common null token
in biological data sets.}

\item{verbose}{Default TRUE, which will report each found file to
the terminal.}
}
\description{
Loads metadata from files accompanying the 'main' data
}
\details{
This package primarily expects matrices to be encoded in
MatrixMarket files that include extensive 'in-line' information
embedded in comments. In some cases it may be desirable to manage
metadata in other files, or you may have file formats other than
MTX that can't hold metadata internally.

In those cases, one or more 'sidecar' files can be included. This
function will expect those files to be named
\code{<BASENAME>-metadata<SOMETHING>}, where <BASENAME> is the full
path of the 'main' data file, and <SOMETHING> is any other text
(including no text).

The metadata files should be TSV tabular and include a header
row. The function will look for an 'id' column - if not found, it
will rename the first column 'id' and use it.

Also supported are files of format
\code{<BASENAME>-parameters<SOMETHING>}, which will be parsed into
a list structure and added as an attribute. These will then be used
to set \code{$param()} values.
}
\examples{

## .makeTempFile() is an internal helper function; Don't use normally!
lolFile <- AnnotatedMatrix:::.makeTempFile("ListOfLists.inp")

## There are two sidecars associated with this file
parseMetadataSidecar( lolFile )

}
