% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{populated}
\alias{populated}
\alias{populatedRows}
\alias{populatedCols}
\title{Populated Rows and Columns}
\arguments{
\item{obj}{Default NULL, which will recover the matrix using
\link{matObj}. Optionally can be a user-provided sparse Matrix.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}

\item{...}{Passed to \link{matObj} (for example, \code{raw})}
}
\value{
A named logical vector, which will be as long as the
    number of rows (for rCounts) or columns (for cCounts). Names
    are taken from the dimension names.
}
\description{
AnnotatedMatrix object methods to flag non-zero rows and columns
}
\details{
\preformatted{
## Method Usage:
myObject$populatedRows( help=TRUE )
myObject$populatedCols( help=TRUE )

myObject$populatedRows( obj=NULL )
myObject$populatedCols( obj=NULL )
}

These methods are just simple wrappers for the
\link[=counts]{counts methods}, which convert
their output to logical using \code{!= 0}. TRUE values indicate at
least one connection to the other dimension.
}
\examples{

s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
# Reset filters, then toss in a metadata filter to make a smaller matrix
s2e$reset()
s2e$filterByMetadata("Description","superfamily")

# What genes and loci still have connections after this?
popRows <- s2e$populatedRows()
popRows[ popRows ]

popCols <- s2e$populatedCols()
s2e$metadata(names(popCols[ popCols ]), c("Symbol", "Description"))

}
\seealso{
\link{counts}, \link{nnZero}
}
