% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileLoaders.R
\name{preCacheMatrixDirectory}
\alias{preCacheMatrixDirectory}
\title{Precache Matrix Directory}
\usage{
preCacheMatrixDirectory(dir = ".")
}
\arguments{
\item{dir}{Default '.', the directory to scan}
}
\value{
A list with two elements; 'pass', the matrix files that
    were succesfully loaded, and 'fail', those that did not.
}
\description{
Attempts to load all MTX files in a directory, generating RDS files
}
\details{
There are two reasons to precache a folder of matrix files; First,
to generate the serialized RDS files in advance, and second, to
scan for files that are malformed and not loading.
}
