% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{print.mapFilter}
\alias{print.mapFilter}
\title{Printing Annotated Matrix filter summarys}
\usage{
\method{print}{mapFilter}(x, color = NULL, ...)
}
\arguments{
\item{x}{The object classed as 'mapFilter'}

\item{color}{Default NULL, which will read the "useCol" attribute
of x. Logical flagging if output should be colorized with
crayon.}

\item{...}{Additional parameters, passed on to message()

s2e  <- AnnotatedMatrix( annotatedMatrixExampleFile() )
s2e$filterSummary( )}
}
\description{
Print method for mapFilter objects generated by
AnnotatedMatrix$filterSummary()
}
\details{
$filterSummary() returns a data frame detailing the number of rows
and columns "zeroed-out" by each applied filter. This method
formats those results into a more human-friendly format.
}
