% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{print.mapResult}
\alias{print.mapResult}
\title{Printing map() results}
\usage{
\method{print}{mapResult}(x, ...)
}
\arguments{
\item{x}{The object classed as 'mapResult'}

\item{...}{Additional parameters, passed on to print.default()

s2e  <- AnnotatedMatrix( annotatedMatrixExampleFile() )
locs <- s2e$map(c("MOT8", "GARS"))
locs}
}
\description{
Print method for mapResult objects generated by AnnotatedMatrix$map()
}
\details{
$map() normally returns a data.frame with no additional class
assigned. If called with format='vector', it will return a named
vector with attributes. The attributes end up being printed by
print.default, which is generally annoying. This method simply
strips attributes before printing.
}
