% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileLoaders.R
\name{rdsIsCurrent}
\alias{rdsIsCurrent}
\title{RDS is Current}
\usage{
rdsIsCurrent(file)
}
\arguments{
\item{file}{Required, the path to the 'main' file (not the RDS)}
}
\value{
\itemize{

  \item TRUE - if the RDS object exists and is newer than all
        supporting files

  \item NA - if the RDS file exists but the original (main) file
        does not

  \item FALSE - All other situations; No RDS, or RDS is older than
        the main file and/or any of the sidecar files

}
}
\description{
Checks if a serialized RDS file is current relative to source files
}
\details{
For speed, matrix files and their supporting annotation files are
serialized to an RDS object. These RDS files will be preferentially
used, unless it is observed (by file dates)
}
