% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{removeEmpty}
\alias{removeEmpty}
\alias{removeEmptyRows}
\alias{removeEmptyCols}
\title{Remove Empty Rows and Columns}
\arguments{
\item{reason}{Default NA, optional human-readable text that will be
included with the \link{setFilters} entry.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
Invisibly, a character vector of removed names
}
\description{
AnnotatedMatrix object methods to remove rows or columns that are all zeroes
}
\details{
\preformatted{
## Method Usage:
myObject$removeEmptyRows( help=TRUE )
myObject$removeEmptyCols( help=TRUE )
myObject$removeEmpty( help=TRUE )

myObject$removeEmptyRows( reason=NA )
myObject$removeEmptyCols( reason=NA )
myObject$removeEmpty( reason=NA )
}

AnnotatedMatrix can work with matrices that have entirely "empty"
rows or columns; For sparse matrices these would correspond to
those that are all zeroes.

There are advantages to leaving such entries in the matrix: In
\code{$map()}, the presence of an empty row/col can be used to
differentiate between "This input no longer has connections" versus
"This input is unrecognized". Additionally, empty entries allow for
name-based access without throwing errors.

However, in some cases it may be desirable to entirely remvove rows
or columns. In particular, crossproducts between matrices requires
the length of the crossed dimension to match.

These operations, like all filters, will only affect the
\link[=matrixUse]{used matrix}. They will not generate new
\link{filterLog} entries, however. Rather, any prior fitlers that
resulted in a row or column becoming fully zeroed-out will have
noted the event. The action will create an entry in
\link{setFilters}, however.

\code{$removeEmpty()} simply wraps both \code{$removeEmptyRows()}
and \code{$removeEmptyCols()}
}
\examples{

s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
# The example matrix has some filters already applied:
rc <- s2e$rCounts()
length(rc)    # Total number of rows = 167
rc[ rc == 0 ] # Rows that no longer connect to columns

gone <- s2e$removeEmptyRows()
gone                 # Removed names
rce <- s2e$rCounts()
length(rce)          # -> 151
rce[ rce == 0 ]      # No more empty rows
}
\seealso{
\link{counts}, \link{nnZero}, \link{populated}
}
