% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{rowcolchanges}
\alias{rowcolchanges}
\alias{rowChanges}
\alias{colChanges}
\title{Changed Row and Column Names}
\value{
A named character vector
}
\description{
Internal AnnotatedMatrix fields tracking row and column names that
needed to be changed when building the object
}
\details{
Row and column names on the matrix must be unique. Some input
formats do not contain this restriction. If any dimension names
need to be altered to force them to be unique, those changes will
be captured here. The values are named character vectors, where the
value of the vector represents the original name, and the names
represent the uniquely-renamed values.

\preformatted{
## THIS FIELD SHOULD NOT BE MODIFIED BY THE USER OR OTHER CODE
## Doing so will probably not cause problems, but I can't conceive
## how it might help anything, either
}
}
\seealso{
\link{rNames}, link{cNames}, link{uniqueNames}
}
