% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{rowcolnames}
\alias{rowcolnames}
\alias{rNames}
\alias{cNames}
\title{Row and Column Names}
\arguments{
\item{new}{Default NULL, which will simply return the current row
or column names. Alternatively can provide a character vector
of new values. This can be used to reorder, remove or add
names. Added names will generate a row or column of all zeros
(no connections). Setting this value is treated as a filter
action, so \link{matrixUse} will be altered and actions noted
in \link{filterLog}.}

\item{raw}{Default FALSE, which will operate on the filtered
matrix. See \link{matObj} for more information. A fatal error
will be thrown if raw is TRUE and new is not NULL (can not
alter raw matrix)}

\item{reason}{Default NA. Only relevant when row is not
NULL. Optional human-readable reason for why the alteration was
made, will be recorded in \link{filterLog} and used to
structure \link{filterSummary}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A character vector of row names
}
\description{
AnnotatedMatrix object method to get/set/reorder the rownames or
colnames of the filtered matrix
}
\details{
\preformatted{
## Method Usage:
myObject$rNames( help=TRUE )

rn <- myObject$rNames( raw=FALSE )
cn <- myObject$cNames( raw=FALSE )

myObject$rNames( newRowNames, reason=NA )
myObject$cNames( newColNames, reason=NA )
}

Normally would be used to simply recover the rownames for the
active matrix. Can also be used to reorder, or to reduce the
accessible names. Unrecognized names will be added to the matrix as
a row or column of all zeros; In most cases this will not be useful
(no connection to other dimension) but might be helpful in some
circumstances.
}
\examples{

s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
length( s2e$rNames() ) # 167 names known
s2e$map('LOC326')      # Maps to six symbols
s2e$rNames(c('AIRE1', 'PGA1', 'AngryPirate', 'AIRE'),
           reason="Nancy asked me to limit analysis to these four genes")
s2e$filterSummary()    # Alteration is recorded in filterLog
length( s2e$rNames() ) # We now have the above four
s2e$map('LOC326')      # We only kept three of the original matches
s2e$removeEmpty()      # Remove rows (and columns) that are only zeros
s2e$matObj()           # A much smaller matrix!

}
\seealso{
\link{reset}, \link{rowChanges}, \link{colChanges}
}
