% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{setFilters}
\alias{setFilters}
\title{Set Filters}
\value{
A character vector
}
\description{
Internal AnnotatedMatrix field holding a character vector
describing filters that have been set
}
\details{
Each operation applied to the matrix to filter out values will add
a new string to this field. The string will be both (somewhat)
human-intelligible, as well as being machine-parsable. The
intention of the field is to allow reapplication of some or all of
the filters, as well as serialize the filters in an ASCII format
that can be stored with output of analyses.

\preformatted{
## THIS FIELD SHOULD NOT BE MODIFIED BY THE USER OR OTHER CODE
## Doing so will probably not cause problems, but I can't conceive
## how it might help anything, either
}
}
\examples{

# The example matrix includes some automatic filters:
s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
s2e$setFilters
}
\seealso{
\link{appliedFilters}, \link{filterSummary}
}
