% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{sharedDimensions}
\alias{sharedDimensions}
\title{Shared Dimensions}
\arguments{
\item{mat2}{Required, an AnnotatedMatrix object that corresponds to
the "right" matrix (again, the matrix object that is being used
to call \code{sharedDimensions} is arbitrarily treated as the
"left" matrix)}

\item{dim1}{Default \code{NULL}, an optional explicitly set "left"
dimension. Can be numeric values 1 or 2 (row or column) or text
values "1" or "2", or text values "row" or "col".}

\item{dim2}{Default \code{NULL}, the same as \code{dim1}, but for
the "right" matrix.}

\item{ignore.case}{Default \code{TRUE}, which will consider overlap
of names irrespective of case.}

\item{warn}{Default \code{TRUE}, which will emit a warning messge
if no overlap was found}

\item{fail.na}{Default \code{TRUE}, which will set both dimensions
to \code{NA} if no overlap was found}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A named integer vector of length 2, each value being either
    \code{1L} (row dimension) or \code{2L} (column dimension), the
    first value being for the "left" matrix, the second for the
    "right" matrix.
}
\description{
Find the most likely shared dimensions between two matrices
}
\details{
\preformatted{
## Method Usage:
myObject$sharedDimensions( help=TRUE )

myObject$sharedDimensions(mat2, dim1=NULL, dim2=NULL, warn=TRUE)
}

Given two matrices (the object calling this method plus a second
one passed to the method), which dimensions on each of the two have
"the most overlap". That is, pick a dimension from the "left"
matrix and a dimension from the "right" matrix that appear to be of
"the same things", based on identifier overlap.

Implemented to support \link{product}, but may have utility
elsewhere.
}
\examples{

mach <- AnnotatedMatrix( annotatedMatrixExampleFile("Machines.ijx") )
comp <- AnnotatedMatrix( annotatedMatrixExampleFile("Components.ijx") )

## Which dimensions appear to be the best overlap?
mach$sharedDimensions(comp)

## You can explicitly define one or both dimensions:
mach$sharedDimensions(comp, dim2='row')

## A warning will be emitted if no overlap is found (either because
## none exists, or you've constrained one or both dimensions to a
## non-overlap)
mach$sharedDimensions(comp, dim1='row')
}
\seealso{
\link{product}
}
