% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileLoaders.R
\name{sidecarFiles}
\alias{sidecarFiles}
\title{Sidecar Files}
\usage{
sidecarFiles(file)
}
\arguments{
\item{file}{Required, the path to the 'main' file}
}
\value{
A character vector of file paths
}
\description{
Given a 'main' file, finds any associated / companion files
}
\details{
Sidecar files are seperate files that carry supporting information
for the main data file. In this package they are used to provide
additional metadata to file formats that can not store such data
internal to the main file
}
\examples{

lolFile <- annotatedMatrixExampleFile("ListOfLists.inp")

sidecarFiles( lolFile )

}
