% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{smallNumberFormatter}
\alias{smallNumberFormatter}
\title{Small Number Formatter}
\usage{
smallNumberFormatter(x)
}
\arguments{
\item{x}{Required, a number, presumably between 0 and 1}
}
\value{
If x == 0, "none". If x >= 0.001, x as a percentage with two
decimal places. Otherwise, x as a fraction with 1 in the numerator,
to 4 significant figures.
}
\description{
Attempt to make small numbers more human-readable
}
\details{
Designed for reporting the fraction of non-zero cells in a sparse matrix
}
