% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{takeLowestThing}
\alias{takeLowestThing}
\title{Take Lowest Thing}
\usage{
takeLowestThing(x)
}
\arguments{
\item{x}{A character vector of strings to pick from}
}
\value{
A single string
}
\description{
From a list of strings, take the lowest/smallest thing
}
\details{
This is a desperation function and should only be used when you
have a character vector and absolutely must reduce it to
"one thing" but you have no real mechanism to do so. It was
designed to address requests to reduce a list of gene accessions
(eg LOC13992, LOC93) to "one gene", but without any other guiding
context. The function will extract the left-most uninterupted
integer that it can find, and then pick the one that is numerically
smallest. The rationale is that "LOC93" was probably annotated
earlier (longer ago) than "LOC13200423", and as such is more likely
to be the "more common" / "higher expressed" / "better annotated" /
"more popular" of the two.
}
\examples{

x <- c("LOC828221", "LOC1234", "LOC39", "HUH?", "LOC39-BETA", "LOC99.243-X")
takeLowestThing(x)

}
\seealso{
\link{map}
}
