% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedMatrix.R
\name{uniqueNames}
\alias{uniqueNames}
\title{Unique Names}
\usage{
uniqueNames(names = character(), rpt = NULL, valid = FALSE,
  verbose = TRUE)
}
\arguments{
\item{names}{Required, a character vector of the names to process}

\item{rpt}{Default NULL. If not NULL, will report to STDERR any
changes that were required. The value of rpt will be used as a
noun in the message (ie if rpt='gene' the message will be of
the form "6 genes required alteration: "...)}

\item{valid}{Default FALSE, whch will utilize make.unique. If TRUE,
then make.names( unique=TRUE) will be used instead}

\item{verbose}{Default TRUE, which will emit a warning if any names
needed to be changed}
}
\value{
A list with two components: "names", the vector of (possibly)
transformed names, and "changes", a named vector where the names
are the new names and the values are the original ones (will only
include altered names)
}
\description{
Wrapper for make.names and make.unique, with reporting of changes
}
\examples{

foo <- c("Apple","Banana","Cherry","Apple")
uniqueNames(foo, rpt="IDs")

}
