% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{annotatedMatrixExampleFile}
\alias{annotatedMatrixExampleFile}
\title{Annotated Matrix Example File}
\usage{
annotatedMatrixExampleFile()
}
\description{
Path to a small example file used for demonstrations
}
\details{
This is just a wrapper around the internal .makeTempFile() function
that pulls up a gene symbol -to-> Entrez Gene ID mapping matrix
(subset of human genes). The file is factorized and has metadata on
both dimensions.
}
