% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{matrixRaw}
\alias{matrixRaw}
\title{Raw Matrix}
\value{
a Matrix:: 'dgTMatrix' sparse matrix
}
\description{
Internal AnnotatedMatrix field holding unaltered (as read from
file) Matrix object
}
\details{
This AnnotatedMatrix object field is populated directly with matrix
data from the source file, and is unaltered. It allows filters to
be cleared using \link{reset}.
}
\seealso{
\link{matObj}, \link{matrixUse}, \link{reset}
}
