% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{matrixUse}
\alias{matrixUse}
\title{Used (Filtered) Matrix}
\value{
a Matrix:: 'dgTMatrix' sparse matrix, or NULL if no filters
    have been applied
}
\description{
Internal field holding the filtered Matrix object
}
\details{
This AnnotatedMatrix object field is \code{NULL} if no filters have
been applied. Otherwise it is a filtered form of
\link{matrixRaw}. You will generally not want to access this field
directly; Instead, use \link{matObj}, as it will by default provide
matrixUse if filters are applied, or will fallback to
matrixRaw. Using \link{reset} will reset all filters by setting
matrixUse to \code{NULL}.
}
\seealso{
\link{matObj}, \link{matrixRaw}, \link{reset}
}
