% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileLoaders.R
\name{parse_ListOfLists_file}
\alias{parse_ListOfLists_file}
\title{Parse List-of-Lists File}
\usage{
parse_ListOfLists_file(file)
}
\arguments{
\item{file}{Required - the path to the LOL file. Files can be
gzipped, provided they have a .gz suffix.}
}
\description{
Parses a List-of-Lists file so it can be read by AnnotatedMatrix
}
\details{
This method is somewhat slow; For many files, you may wish to use
the Perl script "listOfList_to_MatrixMarket.pl" script in the perl/
subdirectory of the AnnotatedMatrix installation.

Once the file is parsed, an RDS version will be written "next to"
it (same path, plus ".rds"), which will load rapidly, unless a
change to the original file triggers an automatic reparsing.
}
