% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{reset}
\alias{reset}
\title{Reset Filters}
\value{
NA, invisibly
}
\description{
Revert any applied filters such that the 'used' matrix is same as 'raw'
}
\details{
AnnotatedMatrix stores two matrices: \link{matrixRaw}
}
\examples{

s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )

## The example matrix has a default filter that suppresses a few symbols:
sketchySymbol <- "B(p51A)"
# No output, though Score=0 (not NA) shows symbol is recognized
s2e$map( sketchySymbol )  

s2e$reset()
s2e$map( sketchySymbol )  # Connection reestablished to LOC8626

}
\seealso{
\link{matObj}, \link{matrixRaw}, \link{matrixUse}
}
